Feat: sns 공유 기능 추가, sns 공유 코드 컴포넌트로 분리
